requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.3
Flask-JWT-Extended==4.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.6.0
PyJWT==2.11.0
Werkzeug==3.1.5
